inference-service/requirements.txt
grpcio==1.74.0
protobuf==6.31.1
PyYAML==6.0.1

# For actual inference (uncomment when needed)
# prometheus-client==0.19.0
torch>=2.0.0
transformers>=4.30.0